game functionnel after new features

+++ src/game/utils/Utils.java
@@ -2,7 +2,7 @@ package game.utils;
 
 public class Utils {
 
-    public static int randInt(double min, double max) {
+    public static int randInt(int min, int max) {
         return (int)(Math.floor(Math.random() * (max - min + 1) + min));
     }
